feat: move events server to dev folder
