+++ contribs/gnodev/pkg/dev/static/hotreload.js
